server/requirements.txt
aws-lambda-powertools
boto3
botocore
fastjsonschema
langchain
langchain-community
langchain-openai
openai
python-dotenv